requirements.txt
absl-py==2.4.0

altair==6.0.0

annotated-doc==0.0.4

annotated-types==0.7.0

anyio==4.12.1

astunparse==1.6.3

attrs==25.4.0

blinker==1.9.0

cachetools==6.2.6

certifi==2026.1.4

charset-normalizer==3.4.4

click==8.3.1

cloudpickle==3.1.2

colorama==0.4.6

contourpy==1.3.3

cycler==0.12.1

fastapi==0.129.0

flatbuffers==25.12.19

fonttools==4.61.1

gast==0.7.0

gitdb==4.0.12

GitPython==3.1.46

google-pasta==0.2.0

grpcio==1.78.0

h11==0.16.0

h5py==3.14.0

idna==3.11

imbalanced-learn==0.14.1

Jinja2==3.1.6

joblib==1.5.3

jsonschema==4.26.0

jsonschema-specifications==2025.9.1

keras==3.13.2

kiwisolver==1.4.9

libclang==18.1.1

llvmlite==0.46.0

markdown-it-py==4.0.0

MarkupSafe==3.0.3

matplotlib==3.10.8

mdurl==0.1.2

ml_dtypes==0.5.4

namex==0.1.0

narwhals==2.16.0

numba==0.63.1

numpy==2.3.5

opt_einsum==3.4.0

optree==0.19.0

packaging==26.0

pandas==2.3.3

pillow==12.1.1


protobuf==6.33.5

pyarrow==23.0.0

pydantic==2.12.5

pydantic_core==2.41.5

pydeck==0.9.1

Pygments==2.19.2

pyparsing==3.3.2

python-dateutil==2.9.0.post0

pytz==2025.2

referencing==0.37.0

requests==2.32.5

rich==14.3.3

rpds-py==0.30.0

scikit-learn==1.8.0

scipy==1.17.0

seaborn==0.13.2

setuptools==82.0.1

shap==0.50.0

six==1.17.0

sklearn-compat==0.1.5

slicer==0.0.8

smmap==5.0.2

starlette==0.52.1


tenacity==9.1.4

tensorflow==2.21.0

termcolor==3.3.0

threadpoolctl==3.6.0

toml==0.10.2

tornado==6.5.4

tqdm==4.67.3

typing-inspection==0.4.2

typing_extensions==4.15.0

tzdata==2025.3

urllib3==2.6.3

uvicorn==0.40.0

watchdog==6.0.0

wheel==0.46.3

wrapt==2.1.2

xgboost==3.2.0

anthropic>=0.18.1
